modernize MuonAnalysis/MomentumScaleCalibration
