+++ MuonAnalysis/MomentumScaleCalibration/plugins/TestResolution.cc
@@ -1,20 +1,95 @@
-#ifndef TESTRESOLUTION_CC
-#define TESTRESOLUTION_CC
+// -*- C++ -*-
+//
+// Package:    TestResolution
+// Class:      TestResolution
+//
+/**\class TestResolution TestResolution.cc MuonAnalysis/MomentumScaleCalibration/plugins/TestResolution.cc
 
-#include "TestResolution.h"
+ Description: <one line class summary>
 
+ Implementation:
+     <Notes on implementation>
+*/
+//
+// Original Author:  Marco De Mattia
+//         Created:  Thu Sep 11 12:16:00 CEST 2008
+//
+//
+
+// system include files
 #include <memory>
+#include <string>
+#include <vector>
 
+// user include files
+#include "DataFormats/Candidate/interface/Candidate.h"
+#include "DataFormats/Candidate/interface/LeafCandidate.h"
+#include "DataFormats/MuonReco/interface/Muon.h"
+#include "DataFormats/MuonReco/interface/MuonFwd.h"
+#include "DataFormats/TrackReco/interface/Track.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "RecoMuon/TrackingTools/interface/MuonPatternRecoDumper.h"
+#include "RecoMuon/TrackingTools/interface/MuonServiceProxy.h"
+
+// For the momentum scale resolution
+#include "MuonAnalysis/MomentumScaleCalibration/interface/ResolutionFunction.h"
+
+// ROOT includes
 #include "TCanvas.h"
 #include "TLegend.h"
+#include "TFile.h"
+#include "TProfile.h"
 
 //
-// constants, enums and typedefs
+// class decleration
 //
 
-//
-// static data member definitions
-//
+class TestResolution : public edm::one::EDAnalyzer<> {
+public:
+  explicit TestResolution(const edm::ParameterSet&);
+  ~TestResolution() override;
+
+private:
+  void analyze(const edm::Event&, const edm::EventSetup&) override;
+  template <typename T>
+  std::vector<reco::LeafCandidate> fillMuonCollection(const std::vector<T>& tracks) {
+    std::vector<reco::LeafCandidate> muons;
+    typename std::vector<T>::const_iterator track;
+    for (track = tracks.begin(); track != tracks.end(); ++track) {
+      // Where 0.011163612 is the squared muon mass.
+      reco::Particle::LorentzVector mu(
+          track->px(), track->py(), track->pz(), sqrt(track->p() * track->p() + 0.011163612));
+      reco::LeafCandidate muon(track->charge(), mu);
+      // Store muon
+      // ----------
+      muons.push_back(muon);
+    }
+    return muons;
+  }
+
+  // ----------member data ---------------------------
+
+  // Collections labels
+  // ------------------
+  edm::InputTag theMuonLabel_;
+  edm::EDGetTokenT<reco::MuonCollection> glbMuonsToken_;
+  edm::EDGetTokenT<reco::TrackCollection> saMuonsToken_;
+  edm::EDGetTokenT<reco::TrackCollection> tracksToken_;
+
+  int theMuonType_;
+  std::string theRootFileName_;
+  TFile* outputFile_;
+
+  TProfile* sigmaPt_;
+
+  int eventCounter_;
+
+  std::unique_ptr<ResolutionFunction> resolutionFunction_;
+};
 
 //
 // constructors and destructor
@@ -99,5 +174,3 @@ void TestResolution::analyze(const edm::Event& iEvent, const edm::EventSetup& iS
 
 //define this as a plug-in
 DEFINE_FWK_MODULE(TestResolution);
-
-#endif  // TESTRESOLUTION_CC
